requirements.txt
playwright>=1.53.0
supabase>=2.16.0
python-dotenv>=1.0.0
